[Modify、Fix] 調整 實裝玩家可以根據選取到的裝備與強化等級來計算能力直
修正 怪物Json檔案 攻擊模式:近戰的內容命名錯誤

+++ game_models.py
@@ -127,7 +127,7 @@ class MonsterDataModel(BasalAttributesDataModel):
     UseSkill: bool = False
     ATK: int = 0
     Hit: int = 0
-    AtkSpeed: int = 0
+    AtkSpeed: float = 0.0
     AttackRange: int = 0
     PursueRange: int = 0
     WalkSpeed: int = 0
